requirements.txt
pytest==8.4.2
allure-pytest==2.15.0
